+++ apps/opik-python-backend/requirements.txt
